ref(web): Consolidate HTTP interface code under `snuba.web`
