+++ snuba/web/static/config.html
