+++ snuba/web/static/css/toggle.css
