+++ snuba/web/static/dashboard.html
