+++ snuba/web/templates/base.html
