+++ tests/test_api.py
@@ -1693,7 +1693,7 @@ class TestCreateSubscriptionApi(BaseApiTest):
     def test(self):
         expected_uuid = uuid.uuid1()
 
-        with patch("snuba.views.uuid1") as uuid4:
+        with patch("snuba.web.views.uuid1") as uuid4:
             uuid4.return_value = expected_uuid
             resp = self.app.post("/subscriptions")
